Renamed physics package to "Math". Added Geometry class
