+++ src/main/java/edu/woodson/Math/Geometry.java
@@ -1,0 +1,4 @@
+package edu.woodson.Math;
+
+public class Geometry {
+}
